Update package versions in package.json, build.gradle, gradle-wrapper.properties, and yarn.lock files
The change includes an update of the dependencies in package.json and devDependencies versions to get the latest features and bug fixes. The gradle version in build.gradle and gradle-wrapper.properties files and several dependencies in yarn.lock were also updated. Moreover, the Kotlin version was upgraded in build.gradle.